refactor(cli): pr.rs を mod.rs + gh.rs に分割 (SPEC-1942 SC-027)
SPEC-1942 SC-027 (各 family file 200-600 行) を pr family で完全達成。 単一
ファイルに偏在していた gh CLI wrapper 群を sub-module に切り離す。

新構成:
- cli/pr/mod.rs (587 行): parse + run + render + tests + gh.rs items の
  re-export (`#[allow(unused_imports)] pub(super) use gh::{...}`)
- cli/pr/gh.rs (625 行): fetch_*_via_gh / create_pr_via_gh / edit_pr_via_gh /
  comment_on_pr_via_gh / fetch_pr_reviews_via_gh / fetch_pr_review_threads_via_gh
  / reply_and_resolve_pr_review_threads_via_gh / fetch_pr_checks_via_gh /
  fetch_pr_review_thread_state_via_gh / parse_pr_checks_items_json /
  parse_pr_checks_items_response / parse_available_fields /
  review_thread_has_comment_body / should_reply_to_review_thread /
  should_resolve_review_thread / extract_pr_url / parse_pr_number_from_url /
  edit_or_create_repo_guard

可視性は pub(crate) (gh.rs items) として、 mod.rs から `pub(super) use`
で family API として再公開。 外部 caller (env.rs / cli.rs / family tests)
は引き続き super::pr::name / crate::cli::pr::name で参照可能、 contract
不変。

cli/pr.rs 1177 行 → mod.rs 587 + gh.rs 625 の 2 ファイルへ分割。 機能・
契約・test 結果は不変 (lib 345 全 pass、 clippy 0 / fmt clean)。
